Reset to last implementation
