--- Assets/Script/Octree/GameManager.cs
@@ -1,39 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class GameManager : MonoBehaviour
-{
-    public GameObject[] worldObjects;
-    public int nodeMinSize = 5;
-    Octree ot;
-    public float minX, maxX, minY, maxY, minZ, maxZ;
-    public float worldBoundary = 10;
-    public static GameManager instance;
-    // Start is called before the first frame update
-    void Start()
-    {
-        instance= this;
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        ot = new Octree(worldObjects, nodeMinSize);
-        minY = ot.boundsD.min.y;
-        maxY = ot.boundsD.max.y;
-        minX = ot.boundsD.min.x;
-        maxX = ot.boundsD.max.x;
-        minZ = ot.boundsD.min.z;
-        maxZ = ot.boundsD.max.z;
-    }
-
-    private void OnDrawGizmos()
-    {
-        if(Application.isPlaying)
-        {
-            Gizmos.color = new Color(0, 1, 0);
-            ot.rootNode.Draw();
-        }
-    }
-}
