--- Assets/Script/Octree/ObjectController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b311da8eab679014999930d228966073
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
