+++ Assets/Sphere.prefab
@@ -12,7 +12,7 @@ GameObject:
   - component: {fileID: 2774370080030858925}
   - component: {fileID: 6828952433265581929}
   - component: {fileID: 2747418434028759965}
-  - component: {fileID: 2158322308285804908}
+  - component: {fileID: 2105993766549509248}
   m_Layer: 0
   m_Name: Sphere
   m_TagString: Untagged
@@ -98,7 +98,7 @@ SphereCollider:
   serializedVersion: 2
   m_Radius: 0.5
   m_Center: {x: 0, y: 0, z: 0}
---- !u!114 &2158322308285804908
+--- !u!114 &2105993766549509248
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
